fix(frontend): get measurements creation working
